interlinked codes + Makefile created

+++ port_scanner/scanner.h
@@ -1,6 +1,7 @@
 #ifndef SCANNER_H
 #define SCANNER_H
+#include <stdint.h>
 
-void scan_ports(const char *ip, int start_port, int end_port);
+void scan_ports(const char *ip, uint16_t start_port, uint16_t end_port);
 
 #endif
